requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.21.0
pillow>=8.3.0
numpy>=1.21.0
matplotlib>=3.4.0

# Web interface
streamlit>=1.28.0

# Configuration and utilities
pyyaml>=6.0
dataclasses-json>=0.5.7

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Optional: For better performance
accelerate>=0.20.0
bitsandbytes>=0.39.0

# Optional: For additional model support
sentencepiece>=0.1.97
protobuf>=3.20.0
